wasm byte code as string for AminoMsg
