+++ __fixtures__/output1/cosmwasm/wasm/v1/tx.amino.ts
@@ -8,7 +8,7 @@ export interface AminoMsgStoreCode extends AminoMsg {
   type: "wasm/MsgStoreCode";
   value: {
     sender: string;
-    wasm_byte_code: Uint8Array;
+    wasm_byte_code: string;
     instantiate_permission: {
       permission: number;
       address: string;
