+++ packages/ast/types/src/encoding/index.d.ts
@@ -2,3 +2,4 @@ export * from './amino';
 export * from './proto';
 export * from './context';
 export * from './object';
+export * from './types';
